Move shift.rs
